+++ deployments/docker-compose.yml
@@ -6,7 +6,7 @@ services:
   app:
     image: golang:latest
     volumes:
-      - .:/go/src/go-with-compose
+      - ../:/go/src/go-with-compose
     working_dir: /go/src/go-with-compose
-    command: go run main.go
-    env_file: config.env
+    command: go run ../cmd/main.go
+    env_file: ../cfg/config.env
